API Changes
#API Endpoints
Added Get and Push, ready to receive data with the front end. Working on delete and put for the pets endpoint.

#HTML Endpoints
Added the index to the '/' endpoint, need to add detail. 

#File structure 
Moved all client side files to the server side in the './webroot/server/public', I deleted the 'client' directory, because there are too many files to upload and it would be redundant to have it, if you need it just pull the 'public' folder.